requirements.txt
google-genai
google-adk
google-cloud-texttospeech
python-dotenv
